pyproject.toml
[tool.poetry]
name = "rest"
version = "0.1.0"
description = ""
authors = ["dan9Protasenia <100715839+dan9Protasenia@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
